requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-appconf==1.0.6
django-ckeditor==6.7.2
django-cleanup==8.1.0
django-debug-toolbar==4.3.0
django-js-asset==3.0.1
fuzzywuzzy==0.18.0
idna==3.6
keyboard==0.13.5
Levenshtein==0.26.1
mysqlclient==2.2.6
pillow==11.0.0
pynput==1.7.6
python-Levenshtein==0.26.1
rapidfuzz==3.9.0
rcssmin==1.1.2
rjsmin==1.2.2
setuptools==68.2.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
wheel==0.41.2
